Update the method "TypeChecker.isString" to a newer assumption
Change the method "TypeChecker.isString" according to a newer assumption that one may want NOT to assume that a string is empty or not. Change unit tests for the method "TypeChecker.isString" according to its latest change. Update the section "Helpful sources" of the README according to the latest change of the method "TypeChecker.isString"